+++ src/search.h
@@ -42,8 +42,8 @@
 typedef struct {
 	char url[MAX_STRING];
 	double speed_start_time;
-	int speed, size;
-	pthread_t speed_thread[1];
+  off_t speed, size;
+  pthread_t speed_thread[1];
 	conf_t *conf;
 } search_t;
